Consolidate acceptance types and start cluster functionality
