+++ test/testutils/acceptance.go
@@ -14,11 +14,25 @@
 package testutils
 
 import (
+	"bytes"
+	"context"
 	"fmt"
 	"net"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"strings"
+	"sync"
+	"syscall"
+	"testing"
 	"time"
 
+	apiclient "github.com/prometheus/alertmanager/api/v2/client"
+	"github.com/prometheus/alertmanager/api/v2/client/general"
 	"github.com/prometheus/alertmanager/api/v2/models"
+
+	httptransport "github.com/go-openapi/runtime/client"
+	"github.com/go-openapi/strfmt"
 )
 
 // AcceptanceOpts defines configuration parameters for an acceptance test.
@@ -69,4 +83,359 @@ func FreeAddress() string {
 	}()
 
 	return l.Addr().String()
+}
+
+// AcceptanceTest provides declarative definition of given inputs and expected
+// output of an Alertmanager setup.
+type AcceptanceTest struct {
+	*testing.T
+
+	opts *AcceptanceOpts
+
+	amc        *AlertmanagerCluster
+	collectors []*Collector
+
+	actions map[float64][]func()
+}
+
+// NewAcceptanceTest returns a new acceptance test with the base time
+// set to the current time.
+func NewAcceptanceTest(t *testing.T, opts *AcceptanceOpts) *AcceptanceTest {
+	test := &AcceptanceTest{
+		T:       t,
+		opts:    opts,
+		actions: map[float64][]func(){},
+	}
+	return test
+}
+
+// Do sets the given function to be executed at the given time.
+func (t *AcceptanceTest) Do(at float64, f func()) {
+	t.actions[at] = append(t.actions[at], f)
+}
+
+// AlertmanagerCluster returns a new AlertmanagerCluster that allows starting a
+// cluster of Alertmanager instances on random ports.
+func (t *AcceptanceTest) AlertmanagerCluster(conf string, size int) *AlertmanagerCluster {
+	amc := AlertmanagerCluster{}
+
+	for range size {
+		am := &Alertmanager{
+			T:    t,
+			Opts: t.opts,
+		}
+
+		dir, err := os.MkdirTemp("", "am_test")
+		if err != nil {
+			t.Fatal(err)
+		}
+		am.dir = dir
+
+		cf, err := os.Create(filepath.Join(dir, "config.yml"))
+		if err != nil {
+			t.Fatal(err)
+		}
+		am.confFile = cf
+		am.UpdateConfig(conf)
+
+		am.APIAddr = FreeAddress()
+		am.ClusterAddr = FreeAddress()
+
+		transport := httptransport.New(am.APIAddr, t.opts.RoutePrefix+"/api/v2/", nil)
+		am.clientV2 = apiclient.New(transport, strfmt.Default)
+
+		amc.ams = append(amc.ams, am)
+	}
+
+	t.amc = &amc
+
+	return &amc
+}
+
+// Collector returns a new collector bound to the test instance.
+func (t *AcceptanceTest) Collector(name string) *Collector {
+	co := NewCollector(t.T, name, t.opts)
+	t.collectors = append(t.collectors, co)
+
+	return co
+}
+
+// Run starts all Alertmanagers and runs queries against them. It then checks
+// whether all expected notifications have arrived at the expected receiver.
+func (t *AcceptanceTest) Run(additionalArgs ...string) {
+	errc := make(chan error)
+
+	for _, am := range t.amc.ams {
+		am.errc = errc
+		t.Cleanup(am.Terminate)
+		t.Cleanup(am.cleanup)
+	}
+
+	err := t.amc.Start(additionalArgs...)
+	if err != nil {
+		t.Log(err)
+		t.Fail()
+		return
+	}
+
+	// Set the reference time right before running the test actions to avoid
+	// test failures due to slow setup of the test environment.
+	t.opts.SetBaseTime(time.Now())
+
+	go t.runActions()
+
+	var latest float64
+	for _, coll := range t.collectors {
+		if l := coll.Latest(); l > latest {
+			latest = l
+		}
+	}
+
+	deadline := t.opts.ExpandTime(latest)
+
+	select {
+	case <-time.After(time.Until(deadline)):
+		// continue
+	case err := <-errc:
+		t.Error(err)
+	}
+}
+
+// runActions performs the stored actions at the defined times.
+func (t *AcceptanceTest) runActions() {
+	var wg sync.WaitGroup
+
+	for at, fs := range t.actions {
+		ts := t.opts.ExpandTime(at)
+		wg.Add(len(fs))
+
+		for _, f := range fs {
+			go func(f func()) {
+				time.Sleep(time.Until(ts))
+				f()
+				wg.Done()
+			}(f)
+		}
+	}
+
+	wg.Wait()
+}
+
+type buffer struct {
+	b   bytes.Buffer
+	mtx sync.Mutex
+}
+
+func (b *buffer) Write(p []byte) (int, error) {
+	b.mtx.Lock()
+	defer b.mtx.Unlock()
+	return b.b.Write(p)
+}
+
+func (b *buffer) String() string {
+	b.mtx.Lock()
+	defer b.mtx.Unlock()
+	return b.b.String()
+}
+
+// Alertmanager encapsulates an Alertmanager process and allows
+// declaring alerts being pushed to it at fixed points in time.
+type Alertmanager struct {
+	T    *AcceptanceTest
+	Opts *AcceptanceOpts
+
+	APIAddr     string
+	ClusterAddr string
+
+	clientV2 *apiclient.AlertmanagerAPI
+	confFile *os.File
+	dir      string
+
+	cmd  *exec.Cmd
+	errc chan<- error
+}
+
+// AlertmanagerCluster represents a group of Alertmanager instances
+// acting as a cluster.
+type AlertmanagerCluster struct {
+	ams []*Alertmanager
+}
+
+// Start the Alertmanager cluster and wait until it is ready to receive.
+func (amc *AlertmanagerCluster) Start(additionalArgs ...string) error {
+	args := additionalArgs
+	for _, am := range amc.ams {
+		args = append(args, "--cluster.peer="+am.ClusterAddr)
+	}
+
+	for _, am := range amc.ams {
+		err := am.Start(args)
+		if err != nil {
+			return fmt.Errorf("failed to start alertmanager cluster: %w", err)
+		}
+	}
+
+	for _, am := range amc.ams {
+		err := am.WaitForCluster(len(amc.ams))
+		if err != nil {
+			return fmt.Errorf("failed to wait for Alertmanager instance %q to join cluster: %w", am.ClusterAddr, err)
+		}
+	}
+
+	return nil
+}
+
+// Members returns the underlying slice of cluster members.
+func (amc *AlertmanagerCluster) Members() []*Alertmanager {
+	return amc.ams
+}
+
+// Start the alertmanager and wait until it is ready to receive.
+func (am *Alertmanager) Start(additionalArg []string) error {
+	am.T.Helper()
+	args := []string{
+		"--config.file", am.confFile.Name(),
+		"--log.level", "debug",
+		"--web.listen-address", am.APIAddr,
+		"--storage.path", am.dir,
+		"--cluster.listen-address", am.ClusterAddr,
+		"--cluster.settle-timeout", "0s",
+	}
+	if len(am.Opts.FeatureFlags) > 0 {
+		args = append(args, "--enable-feature", strings.Join(am.Opts.FeatureFlags, ","))
+	}
+	if am.Opts.RoutePrefix != "" {
+		args = append(args, "--web.route-prefix", am.Opts.RoutePrefix)
+	}
+	args = append(args, additionalArg...)
+
+	cmd := exec.Command("../../../alertmanager", args...)
+
+	if am.cmd == nil {
+		var outb, errb buffer
+		cmd.Stdout = &outb
+		cmd.Stderr = &errb
+	} else {
+		cmd.Stdout = am.cmd.Stdout
+		cmd.Stderr = am.cmd.Stderr
+	}
+	am.cmd = cmd
+
+	if err := am.cmd.Start(); err != nil {
+		return err
+	}
+
+	go func() {
+		if err := am.cmd.Wait(); err != nil {
+			am.errc <- err
+		}
+	}()
+
+	time.Sleep(50 * time.Millisecond)
+	var lastErr error
+	for range 10 {
+		_, lastErr = am.clientV2.General.GetStatus(nil)
+		if lastErr == nil {
+			return nil
+		}
+		time.Sleep(500 * time.Millisecond)
+	}
+	return fmt.Errorf("unable to get a successful response from the Alertmanager: %w", lastErr)
+}
+
+// WaitForCluster waits for the Alertmanager instance to join a cluster with the
+// given size.
+func (am *Alertmanager) WaitForCluster(size int) error {
+	params := general.NewGetStatusParams()
+	params.WithContext(context.Background())
+	var status *general.GetStatusOK
+
+	// Poll for 2s
+	for range 20 {
+		var err error
+		status, err = am.clientV2.General.GetStatus(params)
+		if err != nil {
+			return err
+		}
+
+		if len(status.Payload.Cluster.Peers) == size {
+			return nil
+		}
+		time.Sleep(100 * time.Millisecond)
+	}
+
+	return fmt.Errorf(
+		"expected %v peers, but got %v",
+		size,
+		len(status.Payload.Cluster.Peers),
+	)
+}
+
+// Terminate kills the underlying Alertmanager cluster processes and removes intermediate
+// data.
+func (amc *AlertmanagerCluster) Terminate() {
+	for _, am := range amc.ams {
+		am.Terminate()
+	}
+}
+
+// Terminate kills the underlying Alertmanager process and remove intermediate
+// data.
+func (am *Alertmanager) Terminate() {
+	am.T.Helper()
+	if am.cmd.Process != nil {
+		if err := syscall.Kill(am.cmd.Process.Pid, syscall.SIGTERM); err != nil {
+			am.T.Logf("Error sending SIGTERM to Alertmanager process: %v", err)
+		}
+		am.T.Logf("stdout:\n%v", am.cmd.Stdout)
+		am.T.Logf("stderr:\n%v", am.cmd.Stderr)
+	}
+}
+
+// Reload sends the reloading signal to the Alertmanager instances.
+func (amc *AlertmanagerCluster) Reload() {
+	for _, am := range amc.ams {
+		am.Reload()
+	}
+}
+
+// Reload sends the reloading signal to the Alertmanager process.
+func (am *Alertmanager) Reload() {
+	am.T.Helper()
+	if am.cmd.Process != nil {
+		if err := syscall.Kill(am.cmd.Process.Pid, syscall.SIGHUP); err != nil {
+			am.T.Fatalf("Error sending SIGHUP to Alertmanager process: %v", err)
+		}
+	}
+}
+
+func (am *Alertmanager) cleanup() {
+	am.T.Helper()
+	if err := os.RemoveAll(am.confFile.Name()); err != nil {
+		am.T.Errorf("Error removing test config file %q: %v", am.confFile.Name(), err)
+	}
+}
+
+// UpdateConfig rewrites the configuration file for the Alertmanager cluster. It
+// does not initiate config reloading.
+func (amc *AlertmanagerCluster) UpdateConfig(conf string) {
+	for _, am := range amc.ams {
+		am.UpdateConfig(conf)
+	}
+}
+
+// UpdateConfig rewrites the configuration file for the Alertmanager. It does not
+// initiate config reloading.
+func (am *Alertmanager) UpdateConfig(conf string) {
+	if _, err := am.confFile.WriteString(conf); err != nil {
+		am.T.Fatal(err)
+	}
+	if err := am.confFile.Sync(); err != nil {
+		am.T.Fatal(err)
+	}
+}
+
+// Client returns a client to interact with the API v2 endpoint.
+func (am *Alertmanager) Client() *apiclient.AlertmanagerAPI {
+	return am.clientV2
 }
